+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/brokers/DatacenterBrokerAbstract.java
@@ -87,7 +87,7 @@ public abstract class DatacenterBrokerAbstract extends CloudSimEntity implements
      */
     public DatacenterBrokerAbstract(CloudSim simulation) {
         super(simulation);
-        this.getSimulation().incrementNumberOfUsers();
+        this.getSimulation().incrementNumberOfBrokers();
 
         this.vmsWaitingList = new ArrayList<>();
         this.vmsCreatedList = new ArrayList<>();